requirements.txt
flask
groq
sentence-transformers
chromadb
pytesseract
pillow
deep-translator
langdetect
python-dotenv